backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
anthropic
openai
pydantic
pydantic-settings
python-multipart
slowapi
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
supabase
python-jose[cryptography]
